% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nade.R
\name{clip_bnd}
\alias{clip_bnd}
\title{Clip Bounds}
\usage{
clip_bnd(vec, UB = 1, LB = 0)
}
\arguments{
\item{vec}{vector to clip}

\item{UB}{upper bound}

\item{LB}{lower bound}
}
\description{
Clip bounds of a vector
}
