% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nade.R
\name{long_to_wide_baseR}
\alias{long_to_wide_baseR}
\title{Long to wide}
\usage{
long_to_wide_baseR(sorted_preds)
}
\arguments{
\item{sorted_preds}{sorted predictions}
}
\value{
A data frame of wide predictions
}
\description{
Long to wide
}
