% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nade.R
\name{nade}
\alias{nade}
\title{Run nade}
\usage{
nade(txts)
}
\arguments{
\item{texts}{vector of strings.}
}
\value{
A dataframe with intensities for eight basic emotions.
}
\description{
applies Nade
}
\examples{
hello(c('I love pizza'))
}
