% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nade.R
\name{nadeR_predict}
\alias{nadeR_predict}
\title{Predict emotions}
\usage{
nadeR_predict(model, txts)
}
\arguments{
\item{model}{nadeR model}

\item{txts}{text to predict}
}
\value{
A data frame of emotions
}
\description{
Predict emotions
}
\examples{
model <- nadeR_load()
nadeR_predict(model, "I am happy")

}
