% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nade.R
\name{predict.nade}
\alias{predict.nade}
\title{Predict emotions}
\usage{
\method{predict}{nade}(object, txts)
}
\arguments{
\item{object}{nadeR model}

\item{txts}{text to predict}
}
\value{
A data frame of emotions
}
\description{
Predict emotions
}
\examples{
model <- nadeR_load()
predict(model, "I am happy")

}
