% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nade.R
\name{sort_by_emojis_baseR}
\alias{sort_by_emojis_baseR}
\title{Sort by emojis}
\usage{
sort_by_emojis_baseR(emoji_scores_raw, emoji_lookup)
}
\arguments{
\item{emoji_scores_raw}{raw emoji scores}

\item{emoji_lookup}{emoji lookup table}
}
\value{
A data frame of sorted emojis
}
\description{
Sort by emojis
}
