% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nade.R
\name{nade}
\alias{nade}
\title{Create a New \code{nade} Object}
\usage{
nade()
}
\value{
A new \code{nade} object
}
\description{
Create a new \code{nade} model. The avalable methods are:
\itemize{
\item \code{load} - load the model
}
}
\examples{
model <- nade()
model$load()

}
